Move and rename all forms
