+++ src/views/Sections/Declarations/Edit.vue
