+++ src/views/Sections/Preferences/Edit.vue
